[TUBEMQ-566] manage brokers in region

+++ tubemq-manager/src/main/java/org/apache/tubemq/manager/repository/ClusterRepository.java
@@ -29,7 +29,7 @@ public interface ClusterRepository extends JpaRepository<ClusterEntry, Long> {
      * @param clusterId
      * @return
      */
-    ClusterEntry findClusterEntryByClusterId(Integer clusterId);
+    ClusterEntry findClusterEntryByClusterId(long clusterId);
 
     /**
      * delete cluster by cluster id
