+++ tubemq-manager/src/main/java/org/apache/tubemq/manager/service/ClusterServiceImpl.java
@@ -71,7 +71,7 @@ public class ClusterServiceImpl implements ClusterService {
     }
 
     @Override
-    public ClusterEntry getOneCluster(Integer clusterId) {
+    public ClusterEntry getOneCluster(long clusterId) {
         return clusterRepository
             .findClusterEntryByClusterId(clusterId);
     }
